Modelica: Fix symbol flattening for derived types
This fixes an issue where symbols such as `Current` below are not
flattened properly because they are derived from other types that
eventually inherit from a built-in type. The resulting flat AST
contains, for example, "a.b.__value" symbols that should have instead
had the attributes and modifications from the "__value" symbol copied
into the flat "a.b" symbol. The code to do that is already there, but it
is just not catching the "__value" from "type" symbols.

```
type ElectricCurrent = Real;
type Current = ElectricCurrent;
```