Fix HBridge MODELICAPATH search (first pass)
This fixes name lookup for flattening
Modelica.Electrical.PowerConverters.DCDC.HBridge
(but there are still syntax errors needing work).

MODELICAPATH lookup rules according to spec 3.5 section 13.3
were previously implemented (top-down lookup). However, if name lookup
fails in the parsed AST, then lookup in the file system needs to start
with the bottom-up name lookup rules per section 5.3. Flattening HBridge
shows this issue because it extends Icons.Converter
which is defined in the enclosing PowerConverters package when following
name lookup rules.

This is a first-pass fix because it has only been tested on HBridge and
the code is convoluted since it is added to code that assumed top-down
only file system lookup. A better approach is to revisit the design to
see if it can be better integrated with ast.Class.find_class and use the
tree visitor or listener design patterns for cleaner design while also
making it more efficient. See TODO in the code.