parser/ast: Implement named function arguments
Adds named arguments to `ast.Function` and grammar. In the parser,
function arguments are processed into a list containing positional
arguments and a final item containing a dictionary of named arguments.
In most places, the latter is ignored for backward-compatibility
with current backends.

Grammar is based on Modelica Language Specification 3.5 upon which
the Modelica Standard Library 4.0 is based. Parsing
`Modelica.Blocks.Sources` from that library produced a syntax error
caused by named parameters that is now resolved by this commit.

See https://github.com/modelica/ModelicaSpecification/pull/2378 for a
grammar that parses all of Modelica Standard Library 4.0 without
syntax errors. Some of the changes for this are influenced
by that grammar.