Modelica: Remove connect warning for built-ins
Removes a warning message when connecting built-in types
other than `Real` during flattening. There is no reason why there
should be a message for `Boolean`, `Integer`, or `String` when
there isn't for `Real`.

There was a TODO comment with no explanation just before this
bit of code, so it appears to have been a work in progress that
was not cleaned up after further testing.

I found this during flattening of complicated MSL 4
`Modelica.Electrical.PowerConverters.DCDC.HBridge`
connectors with `Boolean` variables.