WIP: Modelica: Convey `final` through AST
Record the `final` keyword in `ComponentClause` and
`ClassModificationArgument` of the AST and keep the attribute
value during flattening.

This is useful, for instance, for parameters declared final that be
turned into constants in the backend generation of code or when
generating flat components that should not have final parameters show up
in a dialog.

TODO: test cases