Compiler: Move `compiler.py` into `pymoca`
This moves `compiler.py` into the `pymoca` package and adds an entry
point in `setup.py` to create a `pymoca` command line script.

A new `compiler.cli()` function with no arguments is added to conform to
the `setuptools` entry points signature requirement.

In addition to the move of `compiler.py`, there is an API change to
`compiler.main(argv)`. It now expects `argv[0]` to be the program name
(in other words pass in `sys.argv` as-is). This change allows the name
to be passed in from enclosing scripts so that help messages from
argparse appear with the correct program name. Any scripts that import
compiler may need to be updated along with the change in location of the
import.

Existing tests are updated for the new `main(argv)` signature and there
is a new test for the `argv` type checks that were added.