WIP: Implementation Thoughts and Plan
Some initial thoughts on making instantiation and flattening compliant
with Modelica Language Specification version 3.5. As I started looking
into it, I realized that name lookup also needs to be made compliant,
so there are notes on that also as well as an initial plan on steps
to implement it all. See the comments in tree.py for all this.

There is some very initial prototype code for instantiation.
`tree.instantiate_class()` is just a copy/paste stub for now.
Also see the new Instance classes in ast.py.

All of this is definitely a work in progress!