tree: Add extends test showing find_class failure
A case @jackvreeken came up with in review comments of #307. It fails in
Pymoca 0.10 but now passes when using the new `tree.find_name()`.

The test is intended to show the need for "temporary flattening" of the
spec, but find_name is able to correctly handle it using only recursive
lookup without any flattening.