parser: Add short_class_definition modification
Add parsing of modifications on a short class definition, used when the
class replaces another in a redeclare. No previous test cases had this.

Add a test case derived from the recursive instantiation example in the
Modelica Language Specification that we will also use to test
instantiation, but here we will only parse to check this commit.