parser: Add short_class_definition modification
This is used when modifications are applied to the class that is
replacing another in a redeclare. No previous test cases had this.

We add a test case derived from the recursive instantiation example in
the Modelica Language Specification that we will also use to test
instantiation, but here we will only parse to check this commit.