Tree/AST: Change instantiation to follow spec
This implements Modelica class instantiation aligned with the Modelica
Language Specification version 3.5 section 5.6.1. This is a clean-sheet
implementation referencing only the previous Pymoca code and the spec.

Adds a `tree.InstanceTree` class that represents the root of the
instance tree and built-in classes are added to it in `__init__` so that
they are found in name lookup as described in the spec.

Adds `tree.InstanceTree.instantiate()` function that is a recursive
procedure following the spec. Built-in classes act as terminators to the
recursion. The original AST is (hopefully :-) untouched so that it can
be reused in multiple passes of instantiation.

Adds `ast.InstanceSymbol` and `ast.InstanceExtends` to the existing
`ast.InstanceClass` and all three inherit from the new
`ast.InstanceElement` that represents just what is needed for "partial
instantiation" in the spec. A major change from the previous Pymoca code
is that instead of copying extends classes into the local class early in
the instantiation process, separate nodes are created to contain them
("unnamed nodes" in the spec). Only during flattening are they resolved
into the local class.

Several `Exceptions` have been added to better document errors:

- `ModelicaError` is a general exception base class for the rest
- `ModelicaSemanticError`
- `NameLookupError` (previously derived from `Exception`)
- `InstantiationError`
- `UnimplementedError`

Several tests are added:

- `test_intantiation_tree` checks that we have a fully-connected tree
with only instances contained in it.

- `test_inheritance_instantiation` was added earlier for a parse test
only, but now is fully implemented and is the recursive instantiation
example from the instantation section of the spec.

- `test_inheritance_resistor` represents the case that was the original
impetus for this commit and fails with Pymoca 0.10. It is similar to the
Modelica Standard Library v4.0.0
`Modelica.Electrical.Analog.Basic.Resistor` but with some twists to
better test scoping of name lookup during instantiation. This example
goes beyond `test_inheritance_instantiation` which only does top-down
instantiation. This one references classes above in the class tree.

- Many of the existing tests related to instantiation were updated to
test the new code while leaving in the tests of the previous code. Each
has a TODO so we can easily find them when the time comes to update
for the TODO flattening.