WIP: Parser: Implement redeclare within a symbol
Implement redeclare within a symbol, for example:
```Modelica
A a(redeclare B b); // Redeclare a.b as type B
```
In other words, it's a redeclaration within a `component_clause` rule in
the grammar.

Hold multiple symbols in a stack (`list`) during parsing. In the example
above we need to have both `a` and `b` in the stack as we work our way
through the parse rules.

Also eliminate useless code `enterElement_modification` that is not
needed, even before this commit. We were creating extra symbols that
were not used.

TODO: Add test(s) for just redeclare within a symbol.