AST: Move BUILTIN_TYPES from tree.py to ast.py
Fix a problem where name lookup in an instance would not properly find
the `BUILTIN_TYPES` in the scope of a non-instance from the class tree.
We previously had a work-around for this in the tree.py name lookup that
we can now remove.

We also with this change consistently use `self.BUILTIN_TYPES` within
`InstanceTree`.