WIP: Tree: Fix modifications
This is an edit of "Tree: Instantiate using find_name_path"
a28a3775fdf21f72cb1cee12131ed5535b30910e to patch into the
fix-extends-modification branch.

Instances are stored in the passed-in parent instance node or contained
symbol instance, which is often not the parent node from the AST class
definition. The start of the process follows the spec; the top-level
instantiated class parent is the root of the instance tree. (We
do no always add instances to the classes lists according to the AST
structure because we need to allow for multiple instances of the same
name with possibly different modifications.)

We separated out _apply_modifications from _instantiate_partially to
allow it to be used elsewhere. We are currently only using it from
_instantiate_partially, but we keep it separate to reduce function size
and complexity.