parser: Set modification argument scope
Set the scope of modification arguments to the class that contains it
since we know the lexical scope in the parser.

This is the only way to get the correct modification scope in methods of
instantiation that lookup names without saving the path taken to find
the name. This scope should later be set to the instance during
instantiation, according to the Modelica Language spec. This may also be
useful to applications that need the original modification argument
scope before, during, or without instantiation, so we implement it
irrespective of the method of name lookup or instantiation.