tree: Add simple name lookup
Add name lookup for the the "A" part of a name "A.B.C". It tries to
closely follow the Modelica Language Specification v3.5, but there are
several TODO for yet to be implemented features of full name lookup.

Return TypeError if users pass wrong types to the top-level find_name(),
which is common if args are passed in the wrong order, otherwise it will
fail deeper in the code with a less informative error message.

Add a specific `NameLookupError` exception.

Note that we do not prevent code lookup up of names in Modelica code
that would shadow built-in names. The spec says: "The names Real,
Integer, Boolean and String are reserved such that it is illegal to
declare an element with these names.", so we should catch that error in
the parser.

Add tests cases from ModelicaCompliance for "Simple Name Lookup"
as defined in the Modelica Language Specification.