tree: Add import name lookup
Add import name lookup logic that closely follows the previous
implementation. There is a TODO to rewrite this to work with planned
changes to the parser import_clause handler.

Add several import rule checks from the Modelica Language Spec v3.5 that
were not in the previous code.

Add all ModelicaCompliance imported name lookup tests and one global
lookup test. Some are skipped because they need TODO
instantiation/flattening and the global lookup test is skipped because a
grammar update is needed to support it. Finally, one test is skipped due
to a bug in ModelicaCompliance.