parser: Raise ModelicaSyntaxError for ANTLR errors
Change `parse()` and associated functions to raise a
`ModelicaSyntaxError` when the ANTLR parser finds a syntax error.

Add a `ModelicaSyntaxError` exception that inherits from, without
changing, the Python `SyntaxError`. Since it is a `SyntaxError`, an
uncaught `ModelicaSyntaxError` will lead to Python printing a stack
trace with a readable message referencing the Modelica source. We
previously returned `None` from the parser upon syntax errors and ANTLR
would print the syntax error messages. This new way gives users control
of the output.

If `parse_file()` is used and syntax errors are found, the file name
will be added to the raised `ModelicaSyntaxError`.

Add `syntax_error_from_ctx()` to create a `ModelicaSyntaxError` and
populate it with the info from the ANTLR context.

Add `print_syntax_error()` to print the error message from the exception
info.