parser: Prevent declarations of builtin types
Enforce the spec v3.5 section 4.8 which states "The names Real, Integer,
Boolean and String are reserved such that it is illegal to declare an
element with these names."

While it would be nice to just prevent the use of these names in any
identifier ('IDENT' in the grammar), it is more complicated because they
are OK to use in some cases such as the builtin operators `String()` and
`Integer()` (section 3.7.1). So we try to follow the spec and just
prevent the use in the grammar rules `declaration` `class_definition`,
and `short_class_definition`. These also cover `component_clause`,
`element_redeclaration`, and `element_replaceable`.

It makes sense to do this in the parser closer to the source of the
problem where better context can be given using the
`parser.ModelicaSyntaxError` exception that is raised containing
original source with line/column info. It also simplifies later stages
like instantiation by removing the need to deal with this problem.