ast/parser: Add Built-in types to the class tree
Add the built-in types `Real`, `Integer`, `Boolean`, and `String` Symbol
definitions in `Tree` and create the classes that contain them during
`__init__` of the Tree. This is per spec v3.5 section 5.6.1.1 that
states "The builtin classes are put into the unnamed root of the class
tree." See also section 4.8 "Predefined Types and Classes".

We make the built-in class name and symbol name the same for clarity,
though the symbol could be named anything as long as it is used
consistently throughout the code.

Change parser reference to the built-in types to get them from the new
definitions.