ast: Fix node attribute type hints
Fix some type hints for attributes in the AST nodes. Mostly Optional and
List types that were missed. The cause issues with catching errors,
especially as we add new code. There is much yet to be done with typing
in general since there are still many typing errors in the code.