ast: Add Symbol.full_reference() and friends
Add method to Symbol that already existed for Class. One use for this is
to get the global path for exception/error messages.

Add `element_name_tuple` and `element_full_name` functions to return
fully-qualified names as a tuple of strings or a dot-separated string.

This perhaps could use some more design thought in terms of the overall
API.