parser: Add Symbol redeclare
Implement redeclare for a component/symbol, for example:

```Modelica
A a(redeclare B b); // Redeclare a.b as type B
```

In other words, it's a redeclaration within a `component_clause` rule in
the grammar. Previously, we only supported redeclare of a class.

Hold multiple symbols in a stack (`list`) during parsing. In the
example above we need to have both `a` and `b` in the stack as we work
our way through the parse rules.

Fix redeclare symbol within a declaration. To do this, we change the
parser to allow component_clause and component_clause1 from the
grammar to be interleaved. This happens in Modelica when component
redeclares happen within a component declaration. Previously the
parser would get confused because the `component_clause1` was stored
in the same `ASTListener.comp_clause` attribute as the
`component_clause`. We add an additional `.comp_clause1` attribute to
the `ASTListener` to store the `component_clause1` separately.

TODO: Add tests