parser/ast: Track replaceable
Keep track of when both symbols and classes are replaceable so we can
catch during instantiation if they are being redeclared when they are
not replaceable.

Properly track `replaceable` in the parser, including when it also
means `redeclare`. Fix the parser to match the Modelica specification
allowing one of three alternatives in a modification:

    1. `redeclare Real x`
    2. `redeclare replaceable Real x`
    3. `replaceable Real x`

Alternative 3 means the same as alternative 2. All are redeclares of a
symbol. We also fix the similar situation for class redeclares in
modifications.

Rename the existing Symbol `redeclare` attribute to `replaceable`
since the former was not used and symbols/components can be
replaceable without being redeclared. The parser now sets this when
parsing a replaceable component.

Add `replaceable` attribute to `ast.ShortClassDefinition` and
`ast.ComponentClause` so we can fully track replaceable for both
symbols and classes. We keep these two symmetric, though we will only
currently use `ShortClassDefinition.replaceable` during instantiation
of classes. The `replaceable` flag is propagated to the symbols in
`ComponentClause` in the parser and then `Symbol.replaceable` is what
will be used during instantiation for symbols.