tree: Instantiate using find_name_path
Fix instantiation to utilize the new `find_name_path()` and apply
modifications to children in the path.

Store instances in the passed-in parent instance node or contained
symbol instance, which is often not the parent node from the AST class
definition. The start of the process follows the spec; the top-level
instantiated class parent is the root of the instance tree. (We
do no always add instances to the classes lists according to the AST
structure because we need to allow for multiple instances of the same
name with possibly different modifications.)

Instantiate extends clauses included in the name lookup path, including
modifiers.

Add `_instantiate_path_partially()` to replace
`_instantiate_parents_partially()` by building a tree path from parent
to child instead of the other direction. Fully instantiate any extends
clauses in the path and pass down modifications to the children. Use
this to instantiate base classes and symbol types to properly handle
compound names.

Update test_instantiation_tree for the expected TreePath structure.