tree/ast: Change instantiation to follow spec
Change Modelica class instantiation to align more closely with the
Modelica Language Specification version 3.5 section 5.6.1. This is a
clean-sheet implementation referencing only our previous code and the
spec.

Add `tree.instantiate()` to recursively create the class instance
following the procedure in the spec. The top-level instantiated class
parent is the root of the instance tree. Built-in classes act as
terminators to the recursion. The original AST is not modified so that
it can be reused in multiple passes of instantiation.

A major change from the previous code is that instead of copying extends
classes into the local class early in the instantiation process,
separate nodes are created to contain them ("unnamed nodes" in the
spec). This maintains correct scope for name lookup. Only during
flattening are the the unnamed nodes resolved into the local class.

Add a `tree.InstanceTree` class that represents the root of the instance
tree and is created by `tree.instantiate()`. Built-in classes are
instantiated in `__init__` so that they are found in name lookup as
described in the spec. As much as possible, we use the same procedure to
instantiate built-ins as all other classes.

Add `ast.InstanceSymbol` to make symbol instantiation similar to the
existing `ast.InstanceClass`. Both inherit from the new
`ast.InstanceElement` that represents just what is needed for "partial
instantiation" in the spec.

Modify name lookup to look for names in the class tree (AST) if a name
is not found in the instance tree. This detail is our implementation,
not described in the spec. The spec leaves as an implementation choice
how the instance tree and class tree are represented.

Modify name lookup to optionally not look in other inherited classes
when we are doing name lookup for an inherited class. The spec is a
little vague on this, but the compliance test
`ModelicaCompliance\Inheritance\Flattening\InheritedBaseClass.mo` states
that "[...] a name used in an extends-clause may not be inherited".

Add/change exceptions to better document errors:

- `ModelicaError` is a general exception base class for all Modelica
  input code errors.
- `ModelicaSemanticError`
- `NameLookupError` (previously derived from `Exception`, now inherits
  from `ModelicaError`)

Add or modify several tests:

- `test_inheritance_instantiation` is the recursive instantiation
  example from the instantiation section of the spec. (This was added
  earlier for a parse-only test, but now is fully implemented.)

- `test_inheritance_resistor` represents the case that was the original
  impetus for this commit and fails with Pymoca 0.10. It is similar to
  the Modelica Standard Library v4.0.0
  `Modelica.Electrical.Analog.Basic.Resistor` but with some twists to
  better test scoping of name lookup during instantiation. This example
  goes beyond `test_inheritance_instantiation` which only does top-down
  instantiation. This one references classes above in the class tree.

- `test_instantiation_same_names` to check that we can instantiate a
  class with the same name as another in the instance tree with no name
  conflicts.

- `test_instantiation_tree` checks that we have a fully-connected tree
  with only instances contained in it. It also checks parent/child
  relationships in the instantiation tree.

- Several tests check for proper semantic error exceptions for extends
  and redeclare errors.

- `test_nested_symbol_modifier` tests multiple levels of modifications
  on symbols are done properly.

- Many of the existing tests related to instantiation were updated to
  test the new code while leaving in the tests of the previous code.
  Each has a TODO so we can easily find them when the time comes to
  update for the TODO flattening.

- Add helper method `parse_and_instantiate_model()` used in many of the
  new and updated tests to check that the original class tree is not
  modified by instantiation.

Co-authored-by: Jack Vreeken <jack@vreeken.me>