tree: Replace CLASS_SEPARATOR with "."
Remove `CLASS_SEPARATOR = "."` definition` and replace with the literal
`"."` everywhere it is used.

The `CLASS_SEPARATOR` variable was sometimes used and sometimes not, in
both new and relatively old code, pointing to some cognitive overhead.
Just using the literal is shorter and immediately understandable. While
backends might need to replace the class separator, at this stage in the
program flow it is very unlikely it would ever be changed, and even then
find/replace is easy enough.