tree: Set Symbol.class_modification unconditionally
Make all `Symbol.class_modification` be at least an empty
`ClassModification()` to remove the need to check for `None` in
downstream processing. This was already done for `ExtendsClause` so the
two constructs with `.class_modification` are now symmetric.

Remove the now-unnecessary `None` checks in the tree. In one case, also
fix making unnecessary copies of empty `ClassModification`s.